--- Assets/Editor/Tests/Terrain/TerrainSpec.cs
@@ -1,19 +1,0 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using NUnit.Framework;
-using UnityEngine;
-using UnityEngine.TestTools;
-
-namespace Tests
-{
-    public class TerrainSpec
-    {
-        [Test]
-        public void CanGetHeight()
-        {
-            int seed = 0;
-            Terrain terrain = new Terrain(seed,1,1,1);
-            Assert.AreEqual(1, terrain.GetHeight(0,0));
-        }
-    }
-}
