+++ Assets/VRoxel/Terrain/Generator.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 27505727cf626fd45ad636886980d7b4
+guid: e22ab59a81597424684e34f1437d583e
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
